.history/requirements_20251001130109.txt
tensorflow
pandas
gdown
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm

joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .

